feedme/requirements.txt
pip==6.0.8
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.7
django-filter==0.9.2
django-postgrespool==0.3.0
django-redis==3.8.3
djangorestframework==3.1.1
gunicorn==19.3.0
Markdown==2.6.1
psycopg2==2.6
pystache==0.5.4
redis==2.10.3
SQLAlchemy==0.9.9
static==1.1.1
static3==0.5.1
APScheduler==3.0.2
